homework6-zilonggu 1/requirements.txt
Flask==1.1.1
newsapi-python==0.2.6

